+++ packages/react-server/src/entry/react-server.tsx
@@ -6,11 +6,11 @@ import {
 } from "@hiogawa/utils";
 import type { RenderToReadableStreamOptions } from "react-dom/server";
 import reactServerDomServer from "react-server-dom-webpack/server.edge";
-import type {
-  LayoutRequest,
-  ServerLayoutData,
-} from "../features/router/layout-manager";
-import { createLayoutContentRequest } from "../features/router/utils";
+import {
+  type LayoutRequest,
+  type ServerLayoutData,
+  createLayoutContentRequest,
+} from "../features/router/utils";
 import { ejectActionId } from "../features/server-action/utils";
 import { unwrapRscRequest } from "../features/server-component/utils";
 import { createBundlerConfig } from "../features/use-client/react-server";
